Reorganize and update Cypress tests
